backend/requirement.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
dj-database-url==3.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==21.2.0
whitenoise==6.7.0
PyJWT==2.10.1
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
idna==3.11
packaging==25.0
pydantic==2.12.5
python-dateutil==2.9.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
supabase==2.27.2
postgrest==2.27.2
realtime==2.27.2
